Fix build with GCC 8 and yaml-cpp 0.6 (thanks Kevin Kofler)
This adds patches that make the thing build with yaml-cpp 0.6
and GCC 8, the latter submitted upstream (the former is a bit
too hacky to go upstream as-is), and reverts the changes to try
and build with a bundled yaml-cpp. Thanks to Kevin Kofler for
helping to figure out both issues.